Added check for a cluster with the selected group
