+++ src/Components/Clusters/TableOfClusters.js
@@ -1,31 +1,21 @@
-import React, { useEffect, useCallback } from 'react';
+import React, { useEffect } from 'react';
 import { TableContainer, Table, TableHead, TableRow, TableCell, TableBody, IconButton } from '@material-ui/core';
 import DeleteIcon from '@material-ui/icons/Delete';
 import EditIcon from '@material-ui/icons/Edit';
-import { serverURL } from '../Dashboard';
-import { useDispatch, useSelector } from 'react-redux';
-import { showClusters } from '../../Actions/ClusterActions';
+import { useSelector } from 'react-redux';
 
 export function TableOfClusters (props) {
-	const dispatch = useDispatch();
-	const refreshClustersData = useCallback(() => {
-		fetch(`${serverURL}/cluster/list`)
-		.then(response => response.json())
-		.then(data => data.kubernetes_cluster_list)
-		.then(data => dispatch(showClusters(data)))
-		.then(data => console.log(data))
-	}, [dispatch])
 	const VMGroups = useSelector(state => state.vm_group);
 	useEffect(() => {
 		const interval = setInterval(() => {
-			refreshClustersData();
+			props.refreshClustersData();
 		}, 4000);
 		return () => clearInterval(interval);
-	}, [refreshClustersData]);
+	}, [props.refreshClustersData,props]);
 
 	useEffect(() => {
 		setTimeout(() => {
-			refreshClustersData();
+			props.refreshClustersData();
 		}, 100);
 		props.refreshVMGroupData();
 	},[])
